+++ src/main/java/pugz/omni/common/item/OmniBoatItem.java
@@ -41,7 +41,7 @@ public class OmniBoatItem extends BoatItem {
                 Vector3d vec3d1 = playerIn.getEyePosition(1.0F);
 
                 for (Entity entity : list) {
-                    AxisAlignedBB axisalignedbb = entity.getBoundingBox().grow((double) entity.getCollisionBorderSize());
+                    AxisAlignedBB axisalignedbb = entity.getBoundingBox().grow(entity.getCollisionBorderSize());
                     if (axisalignedbb.contains(vec3d1)) {
                         return new ActionResult<>(ActionResultType.PASS, itemstack);
                     }
